Split platform-specific CEF stuff into separate modules
